feat: 검색 로딩 상태 렌더링 추가
