+++ src/dom/render/renderSearchUI.ts
@@ -3,14 +3,12 @@ import { getSearchedMovies } from "../../apis/search/api";
 import TMDBError from "../../TMDBError";
 import { renderResultSectionContent } from "./renderResultSectionContent";
 import { renderThumbnailList } from "./renderThumbnailList";
+import { renderLoadingUI } from "./renderLoadingUI.ts";
 
 export const renderSearchUI = async (keyword: string) => {
   const searchInput = document.getElementById(
     "search-input",
   ) as HTMLInputElement;
-  const banner = document.getElementById("background-container");
-  const resultSection = document.getElementById("result-section");
-  const subTitle = document.getElementById("sub-title");
   const thumbnailListElement = document.getElementById("search-thumbnail-list");
   searchInput.value = keyword;
 
@@ -19,18 +17,16 @@ export const renderSearchUI = async (keyword: string) => {
   let movies: Movie[] = [];
   let errorMessage = "";
 
-  if (!banner || !subTitle || searchInput?.value.trim() === "") return;
+  if (searchInput?.value.trim() === "") return;
 
   try {
+    renderLoadingUI();
+
     const searchResult = await getSearchedMovies({
       query: keyword,
       language: "ko-KR",
       page: 1,
     });
-
-    banner.hidden = true;
-    resultSection?.classList.add("result-section");
-    subTitle.innerText = `"${keyword}" 검색 결과`;
 
     isLastPage = searchResult.page === searchResult.total_pages;
     movies = searchResult.results;
